Put the PATH in .zshenv
.zshrc isn't sourced for non-interactive shells, such as when running
commands over ssh.
So, to set the PATH correctly, set it in .zshenv, which is sourced for
all zsh shells.

Since .zshenv is sourced before the system-wide zsh profile (for login
shells), the PATH variable is often overwritten by the default PATH in
the profile.
So, we source it from .zshrc, and don't add duplicates to the PATH to
workaround this.